+++ grails-app/controllers/io/xh/hoist/impl/XhController.groovy
@@ -20,7 +20,6 @@ import io.xh.hoist.pref.Preference
 import io.xh.hoist.security.AccessAll
 import io.xh.hoist.security.BaseAuthenticationService
 import io.xh.hoist.telemetry.ClientTraceService
-import io.xh.hoist.telemetry.TraceService
 import io.xh.hoist.track.TrackService
 import io.xh.hoist.environment.EnvironmentService
 import io.xh.hoist.user.BaseUserService
@@ -169,11 +168,9 @@ class XhController extends BaseController {
     def submitSpans() {
         ensureClientUsernameMatchesSession()
         def payload = parseRequestJSONArray()
-        clientTraceService.submitClientSpans(payload as List<Map>,)
+        clientTraceService.submitClientSpans(payload as List<Map>)
         renderSuccess()
     }
-
-
 
     //------------------------
     // Json Blobs